[6334] Use a dedicated hook to render temporary edge
Bug: https://github.com/eclipse-sirius/sirius-web/issues/6334
