+++ CHANGELOG.adoc
@@ -60,6 +60,7 @@ It will require contributors to specify the persons involved in the decision pro
 - https://github.com/eclipse-sirius/sirius-web/issues/6047[#6047] [charts] Perform the refresh of the hierarchy representation outside of its representation event processor.
 To perform the refresh of the representation event processor, an instance of `IRepresentationRefresher` will be required.
 A new provisional method `IHierarchyEventProcessor#update` is available to update the state of the hierarchy event processor and publish a new hierarchy to the subscribers.
+- https://github.com/eclipse-sirius/sirius-web/issues/6334[#6334] [diagram] Use a dedicated hook to improve temporary edge rendering
